refactor: migrate communication.rs and prompt_template.rs out of agent module
Source: user
Type: refactor
