+++ src/gateway/session_manager/communication.rs
