+++ src/gateway/session_manager/spawn.rs
@@ -6,7 +6,7 @@
 //! table so `SpawnController` can enforce depth and concurrency limits.
 
 use super::SessionManager;
-use crate::agent::communication::CommunicationConfig;
+use crate::gateway::session_manager::communication::CommunicationConfig;
 use crate::config::agents::ResolvedAgentConfig;
 use crate::gateway::Session;
 use crate::llm::session::ChatSession;
