+++ src/llm/session/mod.rs
@@ -14,7 +14,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, RwLock};
 use tokio_util::sync::CancellationToken;
 
-use crate::agent::communication::CommunicationConfig;
+use crate::gateway::session_manager::communication::CommunicationConfig;
 use crate::llm::session_state::{ChildSessionState, LlmState, ToolExecState};
 use crate::llm::stats::RunningStats;
 use crate::llm::streaming::StreamingSink;
